Rename and move types; implement From traits
